Made some buttons to switch between charts
